feat: add virtual:page-routes/react-router
