+++ packages/vite-glob-routes/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "@tsconfig/strictest/tsconfig.json",
   "include": ["src"],
   "compilerOptions": {
+    "verbatimModuleSyntax": true,
     "noEmit": true,
     "moduleResolution": "Node",
     "module": "ESNext",
