+++ src/server/entry-hattip.ts
@@ -1,11 +1,11 @@
-import apiRoutesHattip from "virtual:api-routes/hattip";
+import globApiRoutes from "virtual:api-routes/hattip";
 import { RequestHandler, compose } from "@hattip/compose";
 import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
 import { TRPC_ENDPOINT } from "../trpc/common";
 import { trpcRouter } from "../trpc/router";
 
 export function createHattipEntry() {
-  return compose(trpcHanlder(), apiRoutesHattip, indexHtmlHanlder());
+  return compose(trpcHanlder(), globApiRoutes, indexHtmlHanlder());
 }
 
 //
